+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog (unofficial)
 
+## [1.3.0] - 2022-06-14
+- Added Move duplicates separate by asset name into <dependencies> folder (right-click on a bundle in AssetBundle Browser Plus)
+- Added Move duplicates by selected and separate by asset name into <dependencies> folder (right-click on multi-bundle in AssetBundle Browser Plus)
+
 ## [1.2.3] - 2022-06-14
 - Optimized LevenshteinDistance algorithm.